requirements.txt
pandas==1.5.3
streamlit==1.18.1
streamlit_option_menu==0.3.2
streamlit-aggrid==0.3.3
ppf-datamatrix==0.1.2